move/#32: 파일 위치 정리
