+++ app/src/main/java/com/example/togedy_android/data/mapper/todata/planner/StudyGoalDateRequestMapper.kt
@@ -1,6 +1,6 @@
 package com.example.togedy_android.data.mapper.todata.planner
 
-import com.example.togedy_android.data.remote.model.request.StudyGoalDateRequestDto
+import com.example.togedy_android.data.remote.model.planner.request.StudyGoalDateRequestDto
 import com.example.togedy_android.domain.model.planner.StudyGoalDate
 
 fun StudyGoalDate.toData() : StudyGoalDateRequestDto = StudyGoalDateRequestDto(
