+++ app/src/main/java/com/example/togedy_android/data/mapper/todata/planner/StudyGoalRequestMapper.kt
@@ -1,6 +1,6 @@
 package com.example.togedy_android.data.mapper.todata.planner
 
-import com.example.togedy_android.data.remote.model.request.StudyGoalRequestDto
+import com.example.togedy_android.data.remote.model.planner.request.StudyGoalRequestDto
 import com.example.togedy_android.domain.model.planner.NewStudyGoal
 
 fun NewStudyGoal.toData(): StudyGoalRequestDto = StudyGoalRequestDto(
