+++ app/src/main/java/com/example/togedy_android/data/mapper/todomain/planner/StudyGoalMapper.kt
@@ -1,6 +1,6 @@
 package com.example.togedy_android.data.mapper.todomain.planner
 
-import com.example.togedy_android.data.remote.model.response.StudyGoalResponseDto
+import com.example.togedy_android.data.remote.model.planner.resposne.StudyGoalResponseDto
 import com.example.togedy_android.domain.model.planner.StudyGoal
 
 fun StudyGoalResponseDto.toDomain(): StudyGoal = StudyGoal(
