+++ app/src/main/java/com/example/togedy_android/data/remote/model/planner/resposne/StudyGoalIdResponseDto.kt
@@ -1,4 +1,4 @@
-package com.example.togedy_android.data.remote.model.response
+package com.example.togedy_android.data.remote.model.planner.resposne
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
